Adding the latest changes to abstract training
